Save change and reset
