+++ src/main/java/com/ezasm/gui/Window.java
@@ -31,6 +31,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.PrintStream;
+import java.util.Timer;
+import java.util.TimerTask;
 
 import static com.ezasm.gui.util.DialogFactory.promptWarningDialog;
 
@@ -385,6 +387,30 @@ public class Window {
      */
     public void handleParseException(Exception e) {
         SystemStreams.printlnCurrentErr(e.getMessage());
+    }
+
+    /**
+     * Auto called Save function to save file periodically
+     *
+     */
+
+    public static void autoSave() {
+        Boolean sw = Window.getInstance().getConfig().getAutoSaveSelected();
+        if (sw){
+            int interval = Window.getInstance().getConfig().getAutoSaveInterval();
+            interval *= 1000;
+            Timer time = new Timer();
+            TimerTask t = new TimerTask(){
+                public void run() {
+                    File fileToUpdate = new File(Window.getInstance().getEditor().getOpenFilePath());
+                    if (fileToUpdate.exists()) {
+                        System.out.println("Save once!");
+                    }
+            
+                }
+            };
+            time.schedule(t, 0, interval);
+        }     
     }
 
 }
