+++ src/main/java/com/ezasm/gui/settings/SettingsPopup.java
@@ -173,6 +173,7 @@ public class SettingsPopup implements IThemeable {
                 instance.config.setAutoSave(instance.AutoSaveButton.getToggleButtonText());
                 instance.config.setAutoSaveInterval(instance.AutoSaveButton.getSliderValue());
                 instance.config.setAutoSaveSelected(instance.AutoSaveButton.getToggleButtonStatus());
+                Window.autoSave();
                 instance.config.saveChanges();
                 instance.applyTheme(new Font(Config.DEFAULT_FONT, Font.PLAIN, instance.config.getFontSize()),
                         EditorTheme.getTheme(instance.config.getTheme()));
@@ -187,6 +188,7 @@ public class SettingsPopup implements IThemeable {
                 instance.AutoSaveButton.setToggleButtonText(Config.DEFAULT_AUTO_SAVE);
                 instance.AutoSaveButton.setToggleButtonStatus(false);
                 instance.AutoSaveButton.setSliderValue(Integer.parseInt(Config.DEFAULT_AUTO_SAVE_INTERVAL));
+                Window.autoSave();
             }
         }
     }
